refactor various helper methods into modules matching respective route paths, e.g. account, recover and providers (sms, email) into their own modules; all tests passing
